Fix issue with tests
